mysq process 修正

+++ src/main/java/red/man10/amanzonkindle/VaultManager.kt
@@ -62,7 +62,7 @@ class VaultManager(private val plugin: JavaPlugin) {
         val resp = economy!!.withdrawPlayer(p, money)
         if (resp.transactionSuccess()) {
             if (p.isOnline) {
-                p.player.sendMessage(ChatColor.YELLOW.toString() + "$" + money.toInt() + "支払いました")
+                p.player.sendMessage(ChatColor.YELLOW.toString() + "$" + money + "支払いました")
             }
             return true
         }
@@ -82,7 +82,7 @@ class VaultManager(private val plugin: JavaPlugin) {
         val resp = economy!!.depositPlayer(p, money)
         if (resp.transactionSuccess()) {
             if (p.isOnline) {
-                p.player.sendMessage(ChatColor.YELLOW.toString() + "$" + money.toInt() + "受取りました")
+                p.player.sendMessage(ChatColor.YELLOW.toString() + "$" + money + "受取りました")
             }
             return true
         }
